+++ website/src/partials/enterprise/EnterpriseFeatures.tsx
@@ -6,11 +6,12 @@ import { Panel } from "../general/Panel";
 import AutoModeIcon from "@site/static/img/icons/auto_mode.svg";
 import SettingsIcon from "@site/static/img/icons/settings.svg";
 import CheckCircleIcon from "@site/static/img/icons/check_circle.svg";
+import { PhoneFrame } from "../PhoneFrame";
+import { BrowserFrame } from "../BrowserFrame";
 
 export function EnterpriseFeatures() {
     return (
         <>
-
 
             <TwoColumns
                 left={<>
@@ -28,16 +29,13 @@ export function EnterpriseFeatures() {
 
                     <p className={"text-xl "}>
                         Every company needs to manage content in different ways.
-                        We
-                        have built FireCMS to be flexible and to be able to
-                        adapt to
-                        different scenarios.
+                        We have built FireCMS to be flexible and customizable.
                     </p>
 
                     <p>
                         Firestore provides a degree of scalability that is hard
-                        to
-                        find in other platforms. It is a great fit for companies
+                        to find in other platforms. It is a great fit for
+                        companies
                         that are looking to build a product that can scale to
                         millions of users. Without the need to worry about the
                         infrastructure.
@@ -93,7 +91,10 @@ export function EnterpriseFeatures() {
                     FireCMS can be extended to
                     fit your needs. We can help you build a custom
                     admin panel that integrates with your infrastructure.
-                    At FireCMS, we believe in <strong>sensible
+                </p>
+
+                <p className={"text-xl "}>
+                    We believe in <strong>sensible
                     defaults</strong>, but we also
                     believe that you should be able to <strong>customize
                     your admin</strong> panel to fit your needs.
@@ -107,50 +108,36 @@ export function EnterpriseFeatures() {
                                 className="max-w-7xl md:max-w-none md:w-full mx-auto md:col-span-9 md:pr-4 md:pr-8 flex justify-center flex-col h-full"
                                 data-aos="fade-right"
                             >
-                                <video
-                                    style={{
-                                        aspectRatio: 842 / 463,
-                                    }}
-                                    className={"rounded-xl border border-solid dark:border-gray-800 border-gray-200"}
-                                    width="100%" loop autoPlay muted>
-                                    <source src="/img/custom_entity_view.mp4"
+                                <BrowserFrame>
+                                    <video
+                                        style={{
+                                            aspectRatio: 842 / 463,
+                                        }}
+                                        width="100%" loop autoPlay muted>
+                                        <source
+                                            src="/img/custom_entity_view.mp4"
                                             type="video/mp4"/>
-                                </video>
+                                    </video>
+                                </BrowserFrame>
                             </div>
 
                             <div
                                 className="md:w-full mx-auto md:col-span-3 md:order-1 my-16 md:my-0 flex justify-center flex-col h-full"
                                 data-aos="fade-left"
                             >
-                                <div className="relative w-72 mx-auto">
-                                    <div
-                                        className="h-6 w-0.5 rounded-l-sm bg-black absolute -left-0.5 top-16"></div>
-                                    <div
-                                        className="h-8 w-0.5 rounded-l-sm bg-black absolute -left-0.5 top-28"></div>
-                                    <div
-                                        className="h-8 w-0.5 rounded-l-sm bg-black absolute -left-0.5 top-40"></div>
-                                    <div
-                                        className="h-auto w-72 bg-black rounded-3xl p-.5 overflow-hidden">
-                                        <div
-                                            className="h-full w-full bg-black rounded-4xl overflow-hidden p-2 relative">
-                                            <div
-                                                className="w-8 h-4 bg-black absolute left-1/2 rounded-b-2xl transform -translate-x-1/2"></div>
-                                            <video
-                                                style={{
-                                                    aspectRatio: 272 / 570,
-                                                    background: "white"
-                                                }}
-                                                className={"max-w-sm rounded-2xl"}
-                                                width="100%" loop autoPlay
-                                                muted>
-                                                <source src="/img/tpa_app.mp4"
-                                                        type="video/mp4"/>
-                                            </video>
-                                        </div>
-                                    </div>
-                                    <div
-                                        className="h-14 w-0.5 rounded-l-sm bg-black absolute -right-0.5 top-32"></div>
-                                </div>
+                                <PhoneFrame>
+                                    <video
+                                        style={{
+                                            aspectRatio: 272 / 570,
+                                            background: "white"
+                                        }}
+                                        className={"max-w-sm rounded-2xl"}
+                                        width="100%" loop autoPlay
+                                        muted>
+                                        <source src="/img/tpa_app.mp4"
+                                                type="video/mp4"/>
+                                    </video>
+                                </PhoneFrame>
                             </div>
                         </div>
                     </div>