+++ website/src/theme/Footer/Layout/index.js
@@ -5,7 +5,6 @@ import { Newsletter } from "../../../partials/general/Newsletter";
 export default function LayoutWrapper(props) {
     return (
         <div className={"mt-16"}>
-            <Newsletter/>
             <section
                 className="relative ">
                 <div
@@ -40,6 +39,7 @@ export default function LayoutWrapper(props) {
                     </div>
                 </div>
             </section>
+            <Newsletter/>
             <Layout {...props} />
         </div>
     );