+++ website/docs/CHANGELOG.md
@@ -2,6 +2,18 @@
 id: changelog
 title: Changelog
 ---
+## [2.0.0-beta.2] - 2023-01-18
+
+### Changed
+- Fixed bug where collection actions were getting their internal state reset.
+- [BREAKING] Removed the prop `loadedEntities` from `CollectionActionsProps`. If
+this prop was useful to you let us know, so we can look for alternative ways to
+provide it.
+- [BREAKING] In the component `EntityCollectionTable`, the prop `ActionsBuilder`
+has been replaced with `actions`.
+- Improved preview of files that are not images, videos, or audio files.
+- Form optimizations
+
 ## [2.0.0-beta.1] - 2023-01-18
 
 This is the first beta release of FireCMS v2.0.0.